Fix the bbb moderator tag not to being found
The ModeratorTagFinder constructs an object with the meetingId that is
setted in the map. But the meetingId being passed was not the same as the one
defined in the map because it was hashed on the front.
So we need to pass the original meetingId (called localMeetingId here) along
with the hashed one (called just meetingId) so that we can look for the tags
in the ModeratorTagFinder.